requirements.txt
anyio==3.6.2
attrs==23.1.0
autoflake==2.1.1
black==23.3.0
certifi==2023.5.7
click==8.1.3
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
isort==5.12.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
openapi-python-client==0.14.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pydantic==1.10.7
pyflakes==3.0.1
python-dateutil==2.8.2
PyYAML==6.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
typer==0.7.0
typing_extensions==4.5.0
